Add cookies.js file in Root to contain cookies,access cookies from cartForPayment and render it on the page
