+++ pages/tracker.js
@@ -91,6 +91,7 @@ function Tracker() {
               <button>To Cart</button>
             </a>
           </Link>
+
           <Link href="/Dogge">
             <a>
               <button className="go-to">Go Pet</button>
